Improve docs for DataFusion integration
