+++ vortex-datafusion/src/persistent/source.rs
@@ -41,9 +41,133 @@ use crate::convert::exprs::ExpressionConvertor;
 use crate::persistent::reader::DefaultVortexReaderFactory;
 use crate::persistent::reader::VortexReaderFactory;
 
-/// Execution plan for reading one or more Vortex files, intended to be consumed by [`DataSourceExec`].
+/// File scan implementation for reading one or more `.vortex` files.
 ///
+/// `VortexSource` is the lower-level read component underneath
+/// [`VortexFormat`]. It is the type DataFusion stores in a [`FileScanConfig`],
+/// and it is ultimately executed through [`DataSourceExec`].
+///
+/// ```text
+///             ▲
+///             │
+///             │  Produce a stream of
+///             │  RecordBatches
+///             │
+/// ┌───────────────────────┐
+/// │     DataSourceExec    │
+/// └───────────────────────┘
+///             ▲
+///             │ uses
+///             │
+/// ┌───────────────────────┐
+/// │      VortexSource     │
+/// └───────────────────────┘
+///             ▲
+///             │ opens `.vortex` files via
+///             │
+///        ObjectStore / VortexReadAt
+/// ```
+///
+/// Most applications reach `VortexSource` indirectly through
+/// [`VortexFormatFactory`]. Use `VortexSource` directly when you are
+/// constructing a `FileScanConfig` yourself or when you need to inject
+/// lower-level behavior such as a custom [`VortexReaderFactory`], an external
+/// [`VortexAccessPlan`], or a specific [`FileMetadataCache`].
+///
+/// # Example
+///
+/// ```rust
+/// use std::sync::Arc;
+///
+/// use arrow_schema::Schema;
+/// use datafusion_datasource::file_scan_config::FileScanConfigBuilder;
+/// use datafusion_datasource::source::DataSourceExec;
+/// use datafusion_datasource::PartitionedFile;
+/// use datafusion_datasource::TableSchema;
+/// use datafusion_execution::object_store::ObjectStoreUrl;
+/// use vortex::VortexSessionDefault;
+/// use vortex::session::VortexSession;
+/// use vortex_datafusion::VortexSource;
+///
+/// let file_schema = Arc::new(Schema::empty());
+/// let source = Arc::new(
+///     VortexSource::new(
+///         TableSchema::from_file_schema(file_schema),
+///         VortexSession::default(),
+///     )
+///     .with_projection_pushdown(true)
+///     .with_scan_concurrency(4),
+/// );
+///
+/// let config = FileScanConfigBuilder::new(ObjectStoreUrl::local_filesystem(), source)
+///     .with_file(PartitionedFile::new("metrics.vortex", 1024))
+///     .build();
+///
+/// let exec = DataSourceExec::from_data_source(config);
+/// # let _ = exec;
+/// ```
+///
+/// # What `VortexSource` Handles
+///
+/// `VortexSource` is responsible for:
+///
+/// - translating DataFusion filters into Vortex predicates when possible,
+/// - retaining the full predicate for file pruning based on statistics and
+///   partition values,
+/// - configuring per-file readers and sharing parsed layout readers across
+///   partitions within the same scan,
+/// - carrying the table schema used for schema evolution and missing-column
+///   adaptation,
+/// - attaching a Vortex metrics registry to the read path.
+///
+/// # Projection And Predicate Behavior
+///
+/// `VortexSource` keeps two related predicate forms:
+///
+/// - `full_predicate`, which is used by DataFusion's `FilePruner` to skip whole
+///   files before they are opened,
+/// - `vortex_predicate`, which contains only the expressions Vortex can evaluate
+///   during the scan.
+///
+/// Projection handling depends on
+/// [`VortexTableOptions::projection_pushdown`]:
+///
+/// - when disabled, `VortexSource` still prunes unreferenced top-level columns,
+///   but DataFusion applies the full projection after the scan,
+/// - when enabled, the scan can evaluate a Vortex-native projection and leave
+///   only unsupported expressions for DataFusion.
+///
+/// # Observability
+///
+/// `VortexSource` owns a Vortex metrics registry for the lifetime of a physical
+/// scan. The registry is passed to the reader and scan builder so I/O and scan
+/// metrics accumulate as the query executes.
+///
+/// Use [`VortexMetricsFinder`] to merge those metrics back into DataFusion
+/// `MetricsSet` values after the plan has run.
+///
+/// # Execution Flow
+///
+/// At execution time:
+///
+/// 1. DataFusion calls `DataSourceExec`, which delegates file opening to
+///    `VortexSource`.
+/// 2. `VortexSource` creates a `VortexOpener` configured with the current
+///    projection, predicate, options, and metrics.
+/// 3. The opener adapts filters and schema for the specific file, applies any
+///    [`VortexAccessPlan`], and builds a Vortex scan.
+/// 4. Scan results are converted into Arrow `RecordBatch` values for
+///    DataFusion.
+///
+/// [`VortexFormat`]: crate::VortexFormat
+/// [`FileScanConfig`]: datafusion_datasource::file_scan_config::FileScanConfig
 /// [`DataSourceExec`]: datafusion_datasource::source::DataSourceExec
+/// [`VortexFormatFactory`]: crate::VortexFormatFactory
+/// [`VortexReaderFactory`]: crate::reader::VortexReaderFactory
+/// [`VortexAccessPlan`]: crate::VortexAccessPlan
+/// [`FileMetadataCache`]: datafusion_execution::cache::cache_manager::FileMetadataCache
+/// [`VortexTableOptions::projection_pushdown`]: crate::VortexTableOptions::projection_pushdown
+/// [`VortexMetricsFinder`]: crate::metrics::VortexMetricsFinder
 #[derive(Clone)]
 pub struct VortexSource {
     pub(crate) session: VortexSession,
@@ -70,10 +194,14 @@ pub struct VortexSource {
 }
 
 impl VortexSource {
-    /// Creates a new VortexSource with default configuration and a provided [`VortexSession`].
-    /// Meant to be use with a [`FileScanConfig`] to scan a file with the provided schema.
+    /// Creates a new `VortexSource` for a table schema and [`VortexSession`].
     ///
-    /// Can be configured using the provided methods.
+    /// The new source starts with:
+    ///
+    /// - all top-level columns projected,
+    /// - no pushed filters,
+    /// - a default Vortex metrics registry,
+    /// - default [`VortexTableOptions`].
     pub fn new(table_schema: TableSchema, session: VortexSession) -> Self {
         let full_schema = table_schema.table_schema();
         let indices = (0..full_schema.fields().len()).collect::<Vec<_>>();
@@ -96,13 +224,21 @@ impl VortexSource {
         }
     }
 
-    /// Enable or disable expression pushdown into the underlying Vortex scan.
+    /// Enables or disables Vortex-native projection evaluation.
+    ///
+    /// This toggles whether `VortexSource` tries to split DataFusion projection
+    /// expressions into a Vortex scan projection plus a leftover DataFusion
+    /// projection.
     pub fn with_projection_pushdown(mut self, enabled: bool) -> Self {
         self.options.projection_pushdown = enabled;
         self
     }
 
-    /// Set a [`ExpressionConvertor`] to control how Datafusion expression should be converted and pushed down.
+    /// Sets the [`ExpressionConvertor`] used to translate DataFusion expressions
+    /// into Vortex expressions.
+    ///
+    /// Override this when the default converter is insufficient for an engine
+    /// integration or for a custom schema-adaptation strategy.
     pub fn with_expression_convertor(
         mut self,
         expr_convertor: Arc<dyn ExpressionConvertor>,
@@ -111,7 +247,10 @@ impl VortexSource {
         self
     }
 
-    /// Set a user-defined factory to create the underlying [`VortexReadAt`]
+    /// Sets a custom factory for the underlying [`VortexReadAt`].
+    ///
+    /// Use this when reads need to go through an application-specific layer
+    /// rather than the default DataFusion [`ObjectStore`].
     ///
     /// [`VortexReadAt`]: vortex::io::VortexReadAt
     pub fn with_vortex_reader_factory(
@@ -122,12 +261,16 @@ impl VortexSource {
         self
     }
 
-    /// Returns the [`MetricsRegistry`] attached to this source.
+    /// Returns the [`MetricsRegistry`] attached to this scan.
+    ///
+    /// The registry is populated as files are opened and scanned. In most
+    /// callers, [`crate::metrics::VortexMetricsFinder`] is the more convenient
+    /// public API for turning the registry contents into DataFusion metrics.
     pub fn metrics_registry(&self) -> &Arc<dyn MetricsRegistry> {
         &self.vx_metrics_registry
     }
 
-    /// Override the file metadata cache
+    /// Overrides the metadata cache used to reuse Vortex footers across scans.
     pub fn with_file_metadata_cache(
         mut self,
         file_metadata_cache: Arc<dyn FileMetadataCache>,
@@ -136,18 +279,20 @@ impl VortexSource {
         self
     }
 
-    /// Set the underlying scan concurrency. This limit is used per Vortex scan operations.
+    /// Sets the per-file Vortex scan concurrency.
+    ///
+    /// This is separate from DataFusion's partition-level parallelism.
     pub fn with_scan_concurrency(mut self, scan_concurrency: usize) -> Self {
         self.options.scan_concurrency = Some(scan_concurrency);
         self
     }
 
-    /// Returns the table options for this source.
+    /// Returns the effective table options for this source.
     pub fn options(&self) -> &VortexTableOptions {
         &self.options
     }
 
-    /// Set the table options for this source.
+    /// Replaces the table options for this source.
     pub fn with_options(mut self, opts: VortexTableOptions) -> Self {
         self.options = opts;
         self
